+++ packages/block-editor/src/components/block-toolbar/index.js
@@ -1,14 +1,7 @@
 /**
- * External dependencies
- */
-import classnames from 'classnames';
-
-/**
  * WordPress dependencies
  */
 import { useSelect } from '@wordpress/data';
-import { Toolbar } from '@wordpress/components';
-import { useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -20,14 +13,12 @@ import BlockSettingsMenu from '../block-settings-menu';
 import BlockSwitcher from '../block-switcher';
 import MultiBlocksSwitcher from '../block-switcher/multi-blocks-switcher';
 import BlockMover from '../block-mover';
-import Inserter from '../inserter';
 
 export default function BlockToolbar( { hasMovers = true } ) {
 	const {
 		blockClientIds,
 		isValid,
 		mode,
-		rootClientId,
 		moverDirection,
 	} = useSelect( ( select ) => {
 		const {
@@ -56,39 +47,10 @@ export default function BlockToolbar( { hasMovers = true } ) {
 			moverDirection: __experimentalMoverDirection,
 		};
 	}, [] );
-	const [ isInserterShown, setIsInserterShown ] = useState( false );
 
 	if ( blockClientIds.length === 0 ) {
 		return null;
 	}
-
-	function onFocus() {
-		setIsInserterShown( true );
-	}
-
-	function onBlur() {
-		setIsInserterShown( false );
-	}
-
-	const inserter = (
-		<Toolbar
-			onFocus={ onFocus }
-			onBlur={ onBlur }
-			// While ideally it would be enough to capture the
-			// bubbling focus event from the Inserter, due to the
-			// characteristics of click focusing of `button`s in
-			// Firefox and Safari, it is not reliable.
-			//
-			// See: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#Clicking_and_focus
-			tabIndex={ -1 }
-			className={ classnames(
-				'block-editor-block-toolbar__inserter',
-				{ 'is-visible': isInserterShown }
-			) }
-		>
-			<Inserter clientId={ blockClientIds[ 0 ] } rootClientId={ rootClientId } />
-		</Toolbar>
-	);
 
 	if ( blockClientIds.length > 1 ) {
 		return (
@@ -99,7 +61,6 @@ export default function BlockToolbar( { hasMovers = true } ) {
 				/> ) }
 				<MultiBlocksSwitcher />
 				<BlockSettingsMenu clientIds={ blockClientIds } />
-				{ inserter }
 			</div>
 		);
 	}
@@ -118,7 +79,6 @@ export default function BlockToolbar( { hasMovers = true } ) {
 				</>
 			) }
 			<BlockSettingsMenu clientIds={ blockClientIds } />
-			{ inserter }
 		</div>
 	);
 }
